+++ src/components/CheckboxGroup/CheckboxGroupLayoutTypes.tsx
@@ -1,0 +1,4 @@
+export enum CheckboxGroupLayoutTypes {
+  Column = "column",
+  Row = "row",
+}