+++ src/components/Form/FormTypes.tsx
@@ -1,8 +1,3 @@
-export enum FormSpacing {
-  ExtraExtraSmall = "extra-extra-small",
-  ExtraSmall = "extra-small",
-  Small = "small",
-  Medium = "medium",
-  Large = "large",
-  ExtraLarge = "extra-large",
-}
+import { GridGaps as FormSpacing } from "../Grid/GridTypes";
+
+export { FormSpacing };