+++ src/theme/components/button.ts
@@ -1,4 +1,4 @@
-// style object for base or default style
+// Style object for base or default style
 const baseStyle = {
   borderRadius: "2px",
   lineHeight: "1.5",
@@ -6,10 +6,14 @@ const baseStyle = {
   cursor: "pointer",
   color: "ui.white",
   justifyContent: "center",
-  py: 2, // var(--space-xs)
-  px: 4, // var(--space-s)
+  py: "xs",
+  px: "s",
   textDecoration: "none",
   wordWrap: "normal",
+  svg: {
+    fill: "currentColor",
+    marginTop: "xxs",
+  },
   _hover: {
     bg: "ui.link.secondary",
   },
@@ -17,89 +21,99 @@ const baseStyle = {
     color: "ui.white",
   },
   _disabled: {
-    bg: "ui.gray.light",
+    bg: "ui.gray.light-cool",
     color: "ui.gray.dark",
     pointerEvents: "none",
     opacity: "1",
   },
 };
-// styles for different sizes ("sm", "md", "lg")
+// Styles for different "lg" size
 const sizes = {
   lg: {
     width: "150px",
   },
 };
-// styles for different visual variants:
-// primary, secondary, link, pill, icon-only
-const variants = {
-  primary: {
-    bg: "ui.link.primary",
-    minWidth: "none",
-    height: "none",
+// Styles for different visual variants:
+// primary, secondary, link, pill, icon-only, callout
+const primary = {
+  bg: "ui.link.primary",
+  minWidth: "none",
+  height: "none",
+};
+const secondary = {
+  bg: "ui.white",
+  border: "1px solid",
+  borderColor: "ui.gray.light-cool",
+  color: "inherit",
+  _hover: {
+    bg: "ui.gray.xx-light-cool",
   },
-  secondary: {
-    bg: "ui.white",
-    border: "1px solid",
-    borderColor: "ui.gray.light",
-    color: "inherit",
-    _hover: {
-      bg: "ui.gray.xxlight",
-    },
-  },
-  link: {
+};
+const link = {
+  bg: "transparent",
+  lineHeight: "2.5",
+  color: "ui.link.primary",
+  textDecoration: "underline",
+  _disabled: {
     bg: "transparent",
-    lineHeight: "2.5",
-    color: "ui.link.primary",
-    textDecoration: "underline",
-    _disabled: {
-      bg: "transparent",
-    },
-    _hover: {
-      bg: "transparent",
-      color: "ui.link.secondary",
-    },
   },
-  pill: {
-    bg: "ui.white",
-    border: "1px solid",
-    borderColor: "ui.gray.light",
-    color: "inherit",
-    borderRadius: "20px",
-    py: 2, // var(--space-xs)
-    paddingInlineStart: 6, // var(--space-m)
-    paddingInlineEnd: 6, // var(--space-m)
-    _hover: {
-      bg: "ui.gray.xxlight",
-    },
+  _hover: {
+    bg: "transparent",
+    color: "ui.link.secondary",
   },
-  "icon-only": {
-    bg: "ui.white",
-    border: "1px solid",
-    borderColor: "ui.gray.light",
-    color: "inherit",
-    _hover: {
-      bg: "ui.gray.xxlight",
-    },
-    paddingInlineStart: 1, // var(--space-xs)
-    paddingInlineEnd: 1, //var(--space-xs)
+};
+const pill = {
+  bg: "ui.white",
+  border: "1px solid",
+  borderColor: "ui.gray.light-cool",
+  color: "inherit",
+  borderRadius: "20px",
+  py: "xs",
+  paddingInlineStart: "m",
+  paddingInlineEnd: "m",
+  _hover: {
+    bg: "ui.gray.xx-light-cool",
   },
-  callout: {
-    bg: "brand.primary",
-    _hover: {
-      bg: "brand.secondary",
-    },
-    _active: {
-      bg: "brand.secondary",
-    },
+};
+const iconOnly = {
+  bg: "ui.white",
+  border: "1px solid",
+  borderColor: "ui.gray.light-cool",
+  color: "inherit",
+  _hover: {
+    bg: "ui.gray.xx-light-cool",
   },
+  paddingInlineStart: "xs",
+  paddingInlineEnd: "xs",
+};
+const callout = {
+  bg: "brand.primary",
+  _hover: {
+    bg: "brand.secondary",
+  },
+  _active: {
+    bg: "brand.secondary",
+  },
+};
+const searchBar = {
+  ...primary,
+  borderLeftRadius: "0",
 };
 
 const Button = {
   baseStyle,
   sizes,
   // Available variants:
-  //   primary, secondary, link, pill, icon-only
-  variants,
+  // primary, secondary, link, pill, icon-only, searchBar
+  variants: {
+    primary,
+    secondary,
+    link,
+    pill,
+    ["icon-only"]: iconOnly,
+    callout,
+    searchBar,
+  },
   // Default values
   defaultProps: {
     size: "md",