+++ src/theme/components/customCheckboxGroup.ts
@@ -1,0 +1,12 @@
+import { checkboxRadioGroupLayout } from "./global";
+
+const CustomCheckboxGroup = {
+  parts: ["legend", "required", "helper"],
+  baseStyle: {
+    ...checkboxRadioGroupLayout,
+  },
+  sizes: {},
+  defaultProps: {},
+};
+
+export default CustomCheckboxGroup;