+++ src/theme/components/label.ts
@@ -1,0 +1,17 @@
+const Label = {
+  parts: ["helper"],
+  baseStyle: {
+    marginBottom: "xs",
+    alignItems: "baseline",
+    display: "flex",
+    fontWeight: "500",
+    justifyContent: "space-between",
+    fontSize: "14px",
+    helper: {
+      fontWeight: "300",
+      fontSize: "-1",
+    },
+  },
+};
+
+export default Label;